docker: do not build parity from source, upgrade parity to v2.1.11
We do not build the parity executable from the sources. Instead the parity
executable is copied directly from the parity/parity docker image.

This gives us a much more deterministic build and reduces built time.

Please not that we can't use the latest parity image, because of the following
issue: https://github.com/paritytech/parity-ethereum/issues/10401