Change equivocation rule to same step
To improve the security of the validators (as
trustlines-network/project#401 states), the rule of the equivocation
proof for having the same block height has been exchanged for the same
block step.
Therefore a global constant has been introduced which numeralize the
step duration. Since this value depends on the chain specification, it
must be adopted in case it changes. In combination with a blocks
timestamp the step can be retrieved.
Furthermore the tests have been extended in a way that they test for the
block step with different timestamp values. By the new introduced
parametrization the total amount of executed cases have increased
drastically.

Related to: #14
See also: trustlines-network/project#401