Rework equivocation test data generation
Instead of using the exported blocks of the Kovan test network and
manipulating them for each test cases within block tuples, a new
generation function for equivocated signed block headers has been
introduced. By the usage of this new module providing that generation
functionality it brings the advantage of less and more clear fixture
definitions, as well as the opportunity to use it as well in
parametrized test cases.

Related to: #26