Optimization of test data generation structure
The previously introduced test data generation has been improved to get
more clean code. This includes renaming, functionality split, as well as
as new parametrization by the fusion of multiple test cases of the same
kind.
Related test modules have been adjusted due to the changes. Furthermore
these modules are using explicitly the timestamp keyword argument for
the block header generation, to show that they are equivocated. This
correlates also with the renaming of the generation function.

Related to: #26