refactor the DepositLocker
remove the deposit function. Each participant can now be registered by the
auction contract when his bid is received. The auction contracts will deposit
the total collected value with one call for all of the participants.

We only keep state of the fact if an address is able to withdraw or not.
The deposited amount is the same for all validators, so there is no need keep
the amount for each validator. Slashing or withdrawing sets this flag to false.

see https://github.com/trustlines-protocol/blockchain/issues/5#issuecomment-482539651