add AuctionStates enum in python test code
This has to be in sync with the declaration in ValidatorAuction.sol. We can use
symbolic names in python, which makes the code easier to understand.

In case we add a new enum or delete an existing one in the solidity code, we
only have to adapt this single enum and not all of the values inside the python
code.