prepare depositing bids into a deposit locker
We introduce a new state 'DepositPending' that is reached when we filled all
slots. Previously filling all slots would have resulted in the state being
'Ended'.

Now, the function depositBids has to be called after all slots are sold. It will
set the new state to 'Ended'.

At the moment we only have a dummy function, that needs to be filled with some
life.

So, this change only introduces the new states and adapts the tests to work with
the new state.