use DepositLocker to lock bids from auction
register bidders with depositor contract when they place their bid, and lock the
total value in the depositBids function, that can be called by anyone.

I'm creating the deposit locker instances inside the fixtures that do deploy the
auction contracts. I'm not happy with that, but we already have some fixtures
that deploy deposit locker contracts, though they are being used for the deposit
locker tests. I would prefer to have separate namespaces for the fixtures.