Extend deposit locker slashing to burn deposit
To have a clear balance of the deposit locker contract, slashed
deposits should not only be locked to withdraw, but also the deposit
amount should be burned to the zero address.

Special case for the testing setup is, that the zero address to burn to
is also the address getting the block rewards and transaction fees. This
must be minded when checking the balances after burning.

Related to: #36
See also: trustlines-network/project#397