add test for close-auction command
This also adds some fixtures that can be used to check that the auction state is
as expected. This can also be used in the other tests, so that we know the
command actually does something. Some of the tests just check that a command has
exits with zero, which doesn't necessarily mean it has done something useful.