Extend validator set contract with historical data
To be able to look up for older validator sets, which where active for
a period of time, the contract has been extend to store the data into
its state. The origin approach to use events on finalizing a new
validator set should be given up, because system calls do not emit.
Therefore each finalized change is pushed to a list with its block
number. Furthermore a mapping of these numbers to the responsive list of
validators has been added.
For the convenience of the user, the ABI has been extend by function to
get the number of finalized sets and their number of addresses.
At last but not least the test contract hast been extended to allow
changing the validator address for test purpose, even without
controlling the system address related to Parity.

Related to: #62