adapt code to compile with solidity 0.5.7
The code changes needed are mostly explicitly setting the data location. I've
followed what the solidity compiler was already suggesting. Breaking changes in
to 0.5.x series are documented in [1].

The problem is, that a lot of tests are failing now, some of them with something
like 'eth.exceptions.InvalidInstruction: Invalid opcode 0x1c @ 18'. My guess
here is that the new default target option is to blame. It changed from
byzantium in 0.5.4 [2] to petersberg in 0.5.7 [3]. My guess is that the default
in 0.4.25 is byzantium.

See https://github.com/trustlines-protocol/blockchain/issues/83

[1] https://solidity.readthedocs.io/en/latest/050-breaking-changes.html
[2] https://solidity.readthedocs.io/en/v0.5.4/using-the-compiler.html#target-options
[3] https://solidity.readthedocs.io/en/v0.5.7/using-the-compiler.html#target-options