Downgrade parity to 2.1.11
We've run into some problems with 2.4.6 and decided to go with the same version
that's running in our current image.

The parity image, which we used to copy the executable from is not available
anymore, so we do download an executable from releases.parity.io.

I've added a check that the parity executable has the right sha256sum. In fact
this is exactly the same executable we've had in our previous release.