upload final image from release branch
We now upload the final image from the release branch. This needs manual
approval on circleci.

After building the image on the release branch, we compare the image with the
latest image from the pre-release branch.

This gives us a way to compare the final release against what we tested before
and makes it transparent by having the thing being rebuilt on circleci.

The show-docker-tree script is written in perl, because perl is installed on the
final image (python is not). It can be called with an docker image name and it will
calls itself inside the docker image via running with 'perl -'.

The script doesn't take modification times into account and in fact they will
differ.

This also enables building of the docker image on all branches. It doesn't take
that long since we do not build parity anymore.