fix networkID in chain spec file
The networkID was previously set to the hexadecimal value 0x4874 inside the
chain spec file and was overwritten to decimal value 4874 inside the user config
file or via command line arguments.

We now set the networkID to hexadecimal '0x130A', which is just 4874. That way
we can get rid of the setting in the user config and can get rid of the
additional --network-id command line argument.