Implement foreign bridge deployment command
Using the home bridge deployment as template, this is the simple extension for
the foreign bridge part. Since this consists of an empty contract without any
dependencies, this is quite simple for the moment. Further adjustments can
become necessary due to future development on the foreign bridge contract.

Related to: #105