Adjust e2e test configuration files
Extend the e2e test chain specifications to be like the origin one. This
includes also the EIPs necessary for the contracts compiled for
petersburg. Also add a balance to the validator account which is used to
deploy the contracts. Even if the validator gets the block reward over
times, this secures enough Ether from the very beginnging.

To deploy the contracts for the e2e tests easily, the shared Parity node
configuration has been extended to unlock an account. Additional shared
parameters have been shifted from the docker-compose description.

Related to: #130
See also: #149