Add init function for block reward contract
To get rid of the fixed coded list of allowed bridge addresses, an
initialize function has been added. This allows to define a single
allowed bridge address once in the beginning. Unless the contract is not
initialized, the add extra receiver call is not allowed. This change
focus on touching as less origin code as possible.

Moreover this opportunity has been used to make the block reward amount
configurable as well. This is very useful for testing purpose.

Related to: #193