Combine and simplify home reward and bridge deploy
The commands to deploy the block reward and bridge contract on the home
chain have been combined. This becomes necessary because the reward
contract must be initialized with the address if the home bridge
contract and vise versa. Therefore a new initialization function for the
reward contract has been added. Both deployments and initialization are
integrated into the deploy-home target.

A new CLI argument has been added for the block reward initialization
argument of the block reward amount.

The CLI argument `--validator-set-address` has been renamed to
`--bride-validators-address` to avoid misunderstandings. The whole
setup consists of a ValdiatorProxy and a BridgeValidators contract.
Since this argument is related to the BridgeValidators contract, the
former naming could be confusing.

Moreover the bridge contract deployment and initialization has been
simplified. Therefore the proxy for the bridge gets its ownership
already transfered within the deployment, as soon as the implementation
links to the deployed bridge contract. Thereby only a single contract
with the address of the proxy and the ABI of the bridge is returned now.
As result the bridge initialization has been simplified without
adjusting the proxy anymore.

Related to: #194
See also: #193