Add configuration setup for bridge e2e tests
For the bridge end-to-end tests, a slightly different/extended setup is
necessary. Therefore an additional docker-compose description, as well
as some environments variables are provided to adapt the default setup.
This includes two validating Parity nodes representing the home and
foreign chain for the tests. To keep it simple, both chains use the same
validator with the same node configuration. The chain specification is
prepared with all EIPs for the petersburg update. Also does the
validator has some initial balance for the later deployment of the
contracts.

Related to: #130