Extend bridge end-to-end tests with deployment
The script does deploy all contracts on the home and foreign chain. The
output of the deployment tools get parsed and forwarded to the next
targets. The home bridge does include a transition to a block reward
now. Afterwards the environment variables for the bridge client services
become adjusted. Token become pre-minted for the validator addresses.

Related to: #130