Add test transfer to bridge end-to-end tests
The script has been extended to check the balance of the validator
address before a transfer and after. In between a token transfer is
transacted to trigger the transfer. The tests wait long enough to make
sure the client and chain have enough time. The test is successful, if
the balance has increased.

Related to: #130