Add tools/quickstart folder
The tools/quickstart folder contains some short python scripts to be
run from the quickstart.sh script. These scripts can be used to import
a private key or a keystore file as a parity account.

We will build and upload a docker image from this folder to dockerhub
and pull this image from the quickstart script if needed.