Use the quickstart image from the quickstart.sh script
This uses the quickstart image in order to import a parity account
from a private key or an existing keyfile. The scripts will read the
password and keys on their own, so we don't do this here anymore.

We now also allow the docker images to be configured via command line
options. This can be used for testing or choosing a different image
e.g. from the master branch.

At the moment we use the trustlines/quickstart:master0. This image has
been uploaded manually. We will switch to an automatically uploaded
image after the first build.