Try to distinguish "wrong password" from "malformed keystore file"
Previously we assumed a ValueError is raised iff the user entered the
wrong password. It could be the case that a ValueError is raised
because the keystore file is malformed.

We now distinguish as good as we can between these two cases.

This now also comes with a test, in case the error being raised does
change when upgrading eth_account.

The circle ci configuration only uses a single job. We may do this for
the other jobs as well where it makes sense. It's much simpler. We
rely on the makefile creating the virtualenv.

Also, we need to take the checksum of the requirements-dev.txt file
into account for the other jobs. This file is being included via '-r
../../requirements-dev.txt' from the other requirements.txt files.

We do this for the new job, but it still has to be changed for the
others.