Feature/bridge end to end tests (#207)
* Add configuration setup for bridge e2e tests

For the bridge end-to-end tests, a slightly different/extended setup is
necessary. Therefore an additional docker-compose description, as well
as some environments variables are provided to adapt the default setup.
This includes two validating Parity nodes representing the home and
foreign chain for the tests. To keep it simple, both chains use the same
validator with the same node configuration. The chain specification is
prepared with all EIPs for the petersburg update. Also does the
validator has some initial balance for the later deployment of the
contracts.

Related to: #130

* Add first simple bridge end-to-end test run script

This script provides the basic functionality to control all chain and
bridge services. It includes correct cleanups between the runs and make
sure depending services are up before needed.
Furthermore a few simple arguments are supported. They are explained
within the local README.

This script doesn't deploy anything so far!

Related to: #130

* Extend bridge end-to-end tests with deployment

The script does deploy all contracts on the home and foreign chain. The
output of the deployment tools get parsed and forwarded to the next
targets. The home bridge does include a transition to a block reward
now. Afterwards the environment variables for the bridge client services
become adjusted. Token become pre-minted for the validator addresses.

Related to: #130

* Add test transfer to bridge end-to-end tests

The script has been extended to check the balance of the validator
address before a transfer and after. In between a token transfer is
transacted to trigger the transfer. The tests wait long enough to make
sure the client and chain have enough time. The test is successful, if
the balance has increased.

Related to: #130

* Use latest deploy tools

* Don't bake chainspec into docker image. This way we can just update the config during runtime.

* Require 8 confirmations for finality

* Set the isValidator mapping in the constructor as well

* Initialize account with more gas, as this is also required for bridge validation

* Install order slightly changed, so the home bridge address is different

* End to end test script WIP
- Add helper functions
- More status output
- Faster waiting

Current Status: Token is successfully transferred and bridge signs transfer transaction, but no new coin is minted. Debug via Parity Log and Traces

* The bridge is actually using the foreign limits on the home bridge. Update deploy tool to reflect that.

* Use latest deploy tools

* Remove REWARD_SOL_FILES as this is part of the normal contracts now.

* blockRewardContractTransition changes based on required installation steps. Should remain stable now.

* Turns out, the transfer worked yesterday evening already, just the calculation of the balance was wrong

* Re-Enable Debug Output

* Install pip, wheel and setuptools first

Otherwise there will be warnings because wheel is missing.

* Use the active virtualenv in run_e2e_test.sh

* Fix cleanup function

We need root rights to clean the data directory. Let's run rm via
docker.

* Fix indentation in ForeigBridge.sol

* Pretty print the shared_key.json

This doesn't hurt anyone, looks better and make the pre-commit
hook (end of line fixer) succeed.