Give user the option to configure the netstats client
We now ask the user if he wants to configure the netstats client.

I've changed the script to create a docker network in order to not
having to use docker's deprecated link functionality.

It uses curl to check if the username/password combination works. Our
webserver has been configured to require a password for the /check/
path.