Extend bridge client config with event fetch start
This allows to configure from which block number on the event fetcher
instances should start to fetch. This differs between the home and
foreign chain.

The event fetcher module has been extended to take this new argument for
the start block number. The foreign chain value is already passed to the
token transfer event fetcher.

Related to: trustlines-network/project#594