Export bridge client contract validation
To avoid redundant code for the validation if given contract address
match to their ABI, this functionality has been exported to a new
module. Thereby the event fetcher could be simplified. The verification
gets applied within the main module now.

This lead to a clean code base again and avoid mixed approaches how to
handle contract to work with.

See this conversation:
https://github.com/trustlines-protocol/blockchain/pull/249/files/da7adf2233532a7f5b3eccc7b591113f9a3505fd#r306206958