Extend bridge client config with event fetch start (#253)
* Extend bridge client config with event fetch start

This allows to configure from which block number on the event fetcher
instances should start to fetch. This differs between the home and
foreign chain.

The event fetcher module has been extended to take this new argument for
the start block number. The foreign chain value is already passed to the
token transfer event fetcher.

Related to: trustlines-network/project#594

* Improve bridge client by pull request review input

This includes naming, shortening of test code and a small bug related to
the argument verification in relation to the actual usage.

Related to: trustlines-network/project#594
See also: #253

* Improve bridge client config option naming.

Relates to: #249

* Fix minor bridge client review comments.

Related to: #253

* Apply requested changes for the bridge client

Related to: #253