Remove obsolete bridge deploy tool module 'utils'
This module where meant to share functionality between the modules for
the home and foreign bridge deployment. This even included more
functionality to include POA contracts as will. Due to the recent
simplification, this is not necessary anymore. Therefore this can be
included into the core module as well.

Related to: #256