Improve bridge client contract validation
The contract validator was only tested for an internal ABI containing
descriptions of type 'event'. For them the contract code does contain
there full signature hash. Function descriptions must be checked for the
first 4 bytes only. The remaining types 'constructor' and 'fallback' are
not supported. In general the whole validation has become more stable.

This has been implemented with the use of the Solidity JSON ABI
description documentation.

Related to: #257