Add account verification for bridge client
To avoid failing transactions during runtime on the home bridge, a new
validation function has been introduced to check the permissions of the
configured signing account. Therefore the home bridge contract needed to
make the validator proxy address public. Thereby it is possible to make
a call to the `isValidator` function of this contract.

Related to: #265