Rework bridge client permission validation concept
The previous approach was to have a single initial check if the
configured account is a bridge validator and thereby has the permission
to confirm transfers. If not the whole client stop.
The improved version now is to preserve the initial check, but only log
a warning to the user. This should be enough to make the decision if
this is intended or not. In the following the validator status is
checked before each confirmation. If it is not given, the user is
informed (again) and the transfer skipped.
This approach allows bridge validators to start their node before they
become active validators and also keep the node running after they have
been removed. The initial misconfiguration check is still there as
useful hint, but not as strong as before.

Related to: #265
See also: #268