Add fund function to home bridge contract
Instead of funding the home bridge contract via a balance assignment in
the genesis block, active funding by a transaction should be possible.
This was already used with the TestHomeBridge contract. By adding
a named payable function called `fund` this is now directly possible at
the default contract. Thereby the extension of the test contract is not
necessary anymore.

Related to: #270