Refactor event fetching
Add fetch_some_events function, which tries to fetch from consecutive
ranges of blocks until it has found some events in a range of blocks
or it has reached the head of the chain.

This fixes multiple issues with the previous implementation:

1. The previous implementation already put events to the queue, but
didn't update its internal state. In the case of an error, there would
have been no way to restart the fetcher.

2. It may have slept for some time even though there could have been
new blocks in the chain to search for events.

The fetch_events_not_seen function will be deleted. It's still used in
the tests at the moment.