Remove redundant code with deploy tools dependency
Due to the history of the deploy tools as extraction from the auction
deploy tool, a huge bunch of code id replicated in both packages.
Therefore the auction deploy tool could be simplified by re-using the
functionality of the deploy tools.
Mark that also the common click CLI options are taken from there. This
means that the associated environment variables are named differently
(`DEPLOY_TOOLS_...`). CLI options specific to the deploy tools keep the
prefix `AUCTION_DEPLOY_`

Related to: trustlines-protocol/contract-deploy-tools#53