Add synthetic event to mark sync status of bridge
For the further processing and to plan the confirmations, the planner
must know when the event fetcher is up to date. Therefore a synthetic
event is put on the queue, which can then be detected and act for.
Because the transfer event queue is still directly connected to the
confirmation sender, this event gets actively ignored for the moment.

Related to: #287