Add CI jobs for bridge validator Docker image
The bridge validator client should be available as Docker image on
DockerHub. Therefore a job to build the image and another to deploy it
have been added to the CircleCI default workflow.
For the moment the image gets only deployed/published on DockerHub when
pushing to the new `release-bridge-validator` branch.

Related to: #301