Simplify Watcher's watch_pending_transactions
We now wait for the transaction to be put on the queue and only call
sleep, when we have to wait for new blocks.

That means that we now call _rpc_latest_block for each
transfer. Previously we didn't do this, but it probably doesn't matter
much.

Some of the tests had to be adapted, since they assume the code is
calling peek, which we don't do anymore.