Add bridge.service
This adds a small class that describes a greenlet to be started. These
are used for what I call 'service greenlets', which are greenlets that
run for the whole lifetime of the program.

We now start them all in the main function, so they all end up in one
pool.

The name attribute will be used for enhanced log output. This will
come with the next commit.