Add initial draft of bridge e2e tests
This adds a first simple end2end test for the bridge, which will just
test that in a setup with all validators online the transfer will be
completed.
However, this also should function as the blueprint for future end2end
tests. Please pay attention to how I tried to make the tests not flaky
by
  - Having up tests for services before proceeding
  - Having timeout tests for fast tests but non flakiness