Allow bridge private key to be a keystore
The `validator_private_key` was meant to expect a hex string encoded
private key. This has been extended to also allow the usage of
keystores. Therefore is has been transformed into a configuration
section or rather a dictionary. Inside it either excepts the old `raw`
definition or a combination of a `keystore_path` with the according
`keystore_password_path`. The raw version takes precedence over the
keystore.

The syntactic verification still happens in the configuration module,
while the semantic one is exported. That is meant to allow reloading of
the configuration itself, but keep the private key static from start.

Related to: #354