Get a new nonce if the nonce is too low
This is part of fixing the nonce handling, see
https://github.com/trustlines-protocol/blockchain/issues/361

Our current way of getting the nonce needs to be adapted to parity (as
described in the issue above). This has not been fixed yet.

But, the fact that we could end up with a nonce that is too low is
still real if the validator signs another transaction.

In that case we actually want to retry to get our transaction on the
chain with a new nonce. This is what this change does.