Ignore Transfer events with zero amount or zero recipient
We filter out Transfer events with zero amount or a zero "from"
address in the transfer recorder. The home bridge contract cannot
handle them anyway.

I would have liked to remove them in the event fetcher. But it's not
as straight forward to do there as one would like.

The tests also had to be adapted in order to provide a Transfer event
providing all fields. Maybe we should get rid of the
get_transfer_hash_event function completely.