Add docker-compose file for quickstart refactoring
In relation to the integration of the bridge client to the quickstart
script, the docker setup gets more interlinked. For example does the
bridge client reuse the validator node to sync with the Trustlines
chain. Also they validators keystore and password file are re-used by
it. After all, a configuration with docker-compose seem to make the
orchestration more easy. It outsource complexity of the script and
simplifies the Docker container handling.

Related to: #302