Extend quickstart tool to handle the full setup
So far the quickstart tool was only an extension for the quickstart
script to handle more complicated tasks like importing a private key
with validation and others.

Due to more and more feature for the quickstart script, which is written
in pure bash, it has become not maintainable and hard to extend.
Therefore the whole logic got transferred to the quickstart tool.
Therefore it handles the complete user interactive flow. Using the click
library makes this task simple to handle. The already existing
functionality to import diverse private key formats could be simply
reused and integrated to the new structure.

Theses changes do also include the integration of the bridge client to
the setup of a validator.

The tool makes use of the recent docker-compose setup. It expects such
in the mounted data directory.

Related to: #371
See also: #302