Extend quickstart bridge configuration for sync
Add the optional start block number fields for the generated bridge
client configuration. Without they are set to zero per default. Because
the foreign as well as the home node need to sync at the beginning, the
bridge client remains in a loop of failures. This is caused by the
attempt to verify the configured contract addresses.

Related to: #302
See also: #378