Fix end2end tests on circle ci
It looks like parity is not mining transactions when running the
end2end tests on circle ci. Passing the --jsonrpc-server-threads=8
command line option seems to fix that.

These machines seem to have 36 CPUs. Maybe parity chooses the default
based on the number of CPUs and we run into a race condition with that
value.