Transfer tokens with the send method
The send method will not revert the transaction if it fails. This can
happen when the receiver is a contract.

If the send fails, we will now still log a TransferCompleted event. An
additional coinTransferSuccessful flag has been added to the event. It
will be true iff the coins have been send successfully.

Transferring the coins to a contract would also have worked previously
with the old code using the transfer method (both send and transfer
use a limit of 2300 gas).

The difference is that we now have the logs on chain, when we
completed a transfer, even when the actual transfer of coins to the
recipient contract has failed.

The bridge doesn't care about this. So, we don't need any new tests
for the bridge.

See https://github.com/trustlines-protocol/blockchain/issues/392