Allow validators to re-check stuck transfers
If the validator set changes it might be the case that some transfers
already have enough confirmations. This change allows a validator to
call confirmTransfer for a Transfer that was already confirmed by
him in order to make the contract complete the Transfer.

See https://github.com/trustlines-protocol/blockchain/issues/389