Do not emit FetcherReachedHeadEvent if node is syncing
We did run into the problem that transfers might be confirmed twice
when starting with a fresh home chain node that is still syncing.

While the actual issue has been that the node would not return the
events when asked for them, I did think of another case where the
event fetcher signals that it has reached the head of the chain, but
the events should still not be processed because the node is syncing.

This commit changes the event handler to not emit the
FetcherReachedHeadEvent if the node is currently syncing.

This doesn't come with any tests, but I did test this with the
configuration of our live test. And indeed we did emit the
FetcherReachedHeadEvent even though the chain was still syncing.

See https://github.com/trustlines-protocol/blockchain/issues/473