Disable warp mode
With active warp mode the parity node will signal that it is finished
with syncing, but when we ask for events it will return an empty list,
because in fact it didn't fully sync.

This resulted in the bridge confirming transfers twice.

This commit changes the docker image to not use warp mode, because I
consider this behaviour a general problem.

See https://github.com/trustlines-protocol/blockchain/issues/473