Make SIGTERM/SIGINT handling work during startup phase
Previously we did receive the signal and tried to kill all greenlets
in the pool. This did miss the main greenlet which might keep on
running some initialization code.

We now run this initialization code inside a dedicated greenlet, which
is part of the main pool and therefore also will be killed, when a
signal is received.

See https://github.com/trustlines-protocol/blockchain/issues/483