Copy docker-compose file to host during quickstart
This copies the docker-compose file to the host during the quickstart
script. This allows to later change paramters to containers and also the
used images.
This makes the quickstart script configurable based on the
docker-compose file used, you can change the name of the install folder,
and you can change the project name used by docker-compose.
To not just simply override user changes, we give him the chance to show
a diff.