Unify parity configuration into a single file
The user-config and valdiator-config files have been merged into one
single template file. The options for the validator are commented out.
All options which need custom values (exclusively addresses) got
placeholders.

The wrapper script does not have a fixed list of additional parity
arguments. They have been integrated into the template file. Adding
custom arguments is still possible.

The wrapper script does not use template snippets anymore, which it add
to the template file. Rather it comment out options in the configuration
file, which are required for the role. Furthermore does it replace the
placeholders for the related option values.

Related to: #493