Downgrade docker-compose file version
To be compatible with the standard Ubuntu 18 version of docker-compose,
which gets installed by the default package repositories, the
configuration version must be changed. Working on a plain Ubuntu 18
machine and install 'docker' and 'docker-compose' lead to an error when
the composition file gets reported. The warning suggest to use either
2.2 or 3.3 instead of 2.4. It turned out that 3.3 is not compatible to
our current configuration, but 2.2 is. Therefore the version in the
docker-compose file has been set to this value.

Related to: #529