Fix Docker image smoke tests due to wrapper script
The new unified parity configuration file/template was setting the
password option per default. This is only necessary if an account gets
unlocked (also for sealing). Therefore the referenced password file is
not available on other scenarios. This happens for the smoke test of the
Docker image as example.

Related to: #493