Rework Parity configuration files
The single configuration file for parity has been split into three, one
for each role the wrapper script knows. Within each file no option is
commented anymore and only for this role relevant options are included.

Moreover have the option for a validator been cleaned up or changed, as
agreed in #546.

Related to: #546