Split parity config files
Split config files to not have to maintain to many config files, as we
need two different ports for both chains and have some docker specific
settings. This sadly breaks the original goal of having an easy way to
see which config is used.
Also the current method of just appending the files will only work as
long as the different configs only change different sections
of the parity config.