Update outdated READMEs and spelling mistakes
There are only a few false information/instructions. So this includes
also a bunch of minor spelling corrections. Only minor adjustments.

The docker-compose setup for the bridge client does not work anymore and
needs a mature update for the configuration as well as documentation.
Therefore issue #608 has been opened to keep it focused.

Related to: trustlines-network/project#798