Remove outdated docker-compose setup for bridge
During the development of the quickstart tool, the documentation and
configuration for the docker-compose setup of the bridge client with two
blockchain nodes has been forgotten. It is not up to date anymore for
a few reasons:
  - the Docker images that are references are not correct anymore
  - the TLBC is live now (it is still stated that it is not)
  - Parity has not proven as stable in light mode, so Geth got decided to prefer
  - the bridge client is not configurable via environment variables anymore

It has been decided that this part of the documentation does not add
much value anymore. It is hard to maintain and is close to what the
quickstart script uses. Moreover does the README in root directory
contain instructions for a manual setup of all components with Docker,
including the bridge.

Related to: #608
See also: trustlines-network/project#798