Add token deposit locker contract
Split the old deposit locker in a base deposit locker contract and a
derived ETH deposit locker contract. This way we can reuse most of the
code for the token deposit locker.
Adapt the most of the deposit locker tests to be run against both ETH
and token deposit locker.