Upgrade pip in the venv
Apparently when running python3 -m venv venv the version of pip in the venv is not the latest that the user has installed globally,  but the version that is bundled with the python version the user is running. In some cases this version is too old and then make install fails. Because of that after initialising venv we upgrade pip